requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==10.0.0
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
psycopg2-binary==2.9.9
django-celery-beat==2.7.0
mercadopago>=1.0.0
celery>=5.3
redis>=5.0
drf-yasg>=1.21
gunicorn>=21.0
requests>=2.31.0
